requirements.txt
Flask
psycopg2-binary
google-cloud-pubsub
python-dotenv